treewide: Update copyright information
While updating the copyright information for the Debian packaging
and thoroughly checking the source files I've noticed that some
copyright information is outdated. I've mostly forgotten to add
myself. Also Andrea deserves to be mentioned as a copyright owner
in the GC about dialog now.
So update the source accordingly.