gui: Allow using GC as regular user
The user can choose whether to use scanmem as root or as regular
user. So make that also possible for GameConqueror by renaming
the script gameconqueror to gameconqueror-root and adding the script
gameconqueror-noroot. Also create a relative symlink gameconqueror
with gameconqueror-root as default target during installation if it
doesn't exist already.

On openSUSE there is no need for root privileges in most cases. It
has been discussed in SUSE bug 985714 that ptrace() will not be
restricted by default on SUSE distributions.
So add the configure option '--disable-root' to create the
gameconqueror symlink with the gameconqueror-noroot target.