autotools: Don't make po/ files without GUI
commit 25d51f1bb9fb4c87a303ec760cfc8fe0d4cb7195 upstream.

For openSUSE it has been decided not to provide a GameConqueror
package together with scanmem. The localization files of
GameConqueror are installed without the configure option
'--enable-gui' set. This is wrong. So don't run make in the po/
directory in that case.

Also restructure configure.ac by moving all GUI-related entries to
the bottom. Unfortunately, intltoolize has to be always called.