Multiple files: Remove the `detect_reverse_change` mechanism
The `detect_reverse_change` is used in INCREASED and DECREASED calls
to also mark variables that changed in the opposite way.

This can be done much more conveniently by using CHANGED, as
`(new > old || new < old) == (new != old)`, apart for very edge
cases like float NaNs

This commit eliminates both the option and the value_t flags and
obtains both a 5% speedup for busy scans and a great simplification
in flag handling, which will be useful later.