scanmem: Move `sm_getcommand()` from `scanmem.h` to `menu.h`
Front-ends are expected to call `sm_execcommand()` directly, so
rename it to `getcommand()` and make it an inner fuction
of the CLI frontend.

Add a `menu.h` header, to be included only by `main.c`, to replace
the `getcommand()` signature that was previsiously in `scanmem.h`.

Move `menu.c` back to GPL, as now it's part of scanmem only.

Also change the required fields in `Makefile.am`.

Reference: #186