Multiple files: Completely rehaul bytearrays handling
Move to a double array of bytes and wildcards, allocated inside the
parsing function. Add a convenience `free_uservalue()`, to free
the memory.

Match the wildcards by masking the incoming memory, so that that
byte always matches the corresponding 0.

This gains a marginal (~5%) speed boost, and will allow several
further optimizations, akin to the string ones.