ptrace: searchregions(): Front-load the whole region even without PROCMEM
Insted of calling `ptrace()` to read each single value,
just read a whole region at once and place it in a buffer,
like it's done with `proc/pid/mem` support.

This leads to a 20-40% reduction in userspace scan time and
provides a common memory interface for both cases, in the form of the
`data` buffer, which will be exploited in the following commits.

Memory consumption of the !PROCMEM case is raised, though, at the
levels of the PROCMEM case.