libscanmem: Completely rewrite incoming memory management
WARNING: This commit is BIG, but it's the minimal set of changes
that compiles AND maintains all the existing functionality.

Rationale:
* The natural type for incoming memory is a pointer into a flat memory
  array, rather that an 8byte chunk, which becomes really difficult
  to use when scanning for VLT
* The natural 'flags' for incoming memory is just its length, not
  a set of values it might be

As such, move from a `value_t` to a `mem64_t* + memlength` approach,
updating accordingly all the related functions.

The raw memory is not copied out by default, this would expose the
program to unaligned access problems on arches that don't support it.
The problem is solved by the use of a packed struct, `mem64_t`.

`checkmatches()` needs to use the same scan_routine interface,
so the `peekbuf` cache is repurposed as a full mirror of the
process memory, with mem64_t* pointing into it.

The `peekbuf` cache has been expanded to fit 2^16 values (plus a
bit of manouvering space), which is the maximum allowed length
for VLT.

`sm_peekdata()` has been changed to return the new `mem64_t*`
interface and has been tweaked for speed, via the
`extern inline` declaration and the use of `memmove()`.

The old matched memory is stored in a way that makes a simple mem64_t*
indexing impossible, so it's still converted to `value_t` on the fly.

`flags_to_memlength()` has been moved to `ptrace.c`, as it's
now called in the hot path and needs to be inlined.

Endianness support has been completely reworked, as now we avoid
copying the incoming memory in general, so it's been integrated into
the scan function themselves.
This shifts the choice to `sm_choose_scanroutine()` time, so
endianness support has no performance penalty in the normal case.
This also allows to fine-tune which scans support endianness and which
doesn't, currently support is incomplete, but all the supported
scans are correct.

Advantages:
* Large VLT now scan without calling peekdata, for absolutely massive
  scan time improvements (2x-100x)
* VLT scans can be fully length-checked, avoiding corner cases
* `searchregions()` is 10-40% faster, depending on match density.
* `checkmatches()` is slightly faster in the sparse memory case
* Endianness swaps are now performed only if needed, fixing several
  bugs with swapping VLT parts