handlers: handler__string(): Allocate a copy of the input string
Instead of pointing the uservalue's string_value to the middle
of the commandline string (a pointer aligned at a 2 bytes
boundary), allocate a new copy with `malloc()`.

This way the target string is aligned by `malloc()` at an alignment
high enough for every type, which avoids unaligned access when
the scan routines will read it as an uint64 stream.

Also rework a bit the whole handler and add a check for
the maximum length that we can scan (UINT16_MAX).

Reported by: Sebastian Parschauer (@sriemer)
Reference: #259