libscanmem: Replace match_flags' bitfield with an enum
Instead of using an union of a bitfield and an integer, use
a simpler enum to emulate bit access.

Rationale:
* No more implementation-definedness in the bitfield implementation,
  so that we can guarantee a reinterpretation as length
  that's correct in all cases and endians
* Avoids to continuously type-pun the stored value between
  bitfield and uint16_t (for flags-wide bitwise operations),
  which avoids a lot of strange fringe behaviours
* Simple integers have more guarantees and are way easier
  for the compiler to optimize and for the programmer to
  work with

Note: the `zero_match_flags()` function was removed, as it's
now unnecessary, assign `flag_empty` to the variable instead.

Snapshot time (dense scan) is improved by a further 25%.
Non dense scan are slightly faster as well (2-5%).