value.c: Make `parse_uservalue_int()` work well with ints in [2^63,2^64)
Parse the input string both as int64 and uint64 and load flags
and values based on the (possibly invalid) parsed values.

Now that each uservalue is guaranteed to have at least one of the
64 bits flags set (as opposed to always both), teach
`parse_uservalue_number()` to check the flags before filling
the floats.

Also import the same detailed bounds checking to
`parse_uservalue_number()`, so that huge floats (>2^64)
can't pass for int64's.

Closes: #234