targetmem: add_element(): Do not zero OVaMI slot that will be overwritten
In the code path that extends the swath, the intervening bytes need
to be zerored. Current code zeroed the last place as well, and
then proceeded to overwrite it with the incoming match.

Change it so that it doesn't zero said last place.

Also improve readability of nearby branch.

Snapshot time down 5% on my hardware.