ptrace: Add procmem support for `sm_checkmatches()`
Use `pread()` on the `proc/pid/mem` to fill the `peekbuf` cache,
instead of `ptrace()`, where it's supported.

This allows reading more data for each syscall and cutting
a huge portion of them (~99%).

Introduce a general `readmemory()` function to abstract away
the difference in the two cases.
Also use it for the other memory reading operations, like in
`searchregions()` and `sm_read_array()`.

Change the signature of `sm_read_array()` to avoid casts with
the new function and change the signature of the related
`sm_write_array()` as well to match.

Performance summary:
* Memory is read in chunks of 2048 bytes, using larger chunks
  does not help dense scans further
* Dense checkmatches time is dramatically decreased, around 3-4x,
  as the bottleneck was starting each syscall
* Sparse checkmatches time changed as follows:
  * For values not found in the cache, `pread` is ~30% slower than
    `ptrace`, this could be mitigated by smaller chucks, but it
    cannot be brought below 10% even with chunk size below 8
  * Therefore, relative timing depends on "cache occupancy" of
    each chunk (2048 bytes),
    at 1.3 values per chunk on average they break even