ptrace: Micro-optimize the comparison expressions in `sm_peekdata()`
Replace the expressions used for the cache decision with
mathematically equivalent ones (as overflow is impossible in this case),
but that the compiler can optimize better.

On my system (gcc-7, amd64) this saves 10% time in a `snapshot;1` scan.

This optimization is really unstable, as equally simple and functionally
equivalent changes lead to slower code, so those ugly expressions
could maybe be rewritten better once compilers get smarter.

Also sneak in a trivial fix for a71e1520, as `readmemory()`
returns 0 on failure, not -1.