ptrace: Return to the old comparison expressions in `sm_peekdata()`
Commit 2bcbe85 "ptrace: Micro-optimize the comparison
expressions in `sm_peekdata()`" changed said expressions with
ones in which the pointer difference between `reqaddr` and
`peekbuf.base` is held in a temporary signed variable.

In 32 bit systems this variable can overflow (as pointer differences
can be bigger than 2^31), leading to crashes.
The old code performed the comparison in a way that's always safe.

Many (correct) variations of the expressions were tried, but
all of them regressed on speed in at least one of the two arches.
So just return to the old implementation.

Tested on ARM 32bit and amd64.

Fixes: 2bcbe859