handlers: set: Introduce and use sm_process_is_dead()
For setting values continuously in target memory it is annoying to
see the following set handler error once the target process
terminates:
error: failed to attach to 7956, No such process
error: failed to set a value.
Please enter current value, or "help" for other commands.

A function to detect if a process is running, a zombie, or dead is
required. The zombie and the error case should be treated as dead
so that we see the unlikely case that the kernel interface changes.
The required check_process() function has been implemented for
ugtrain long time ago already and optimized recently. It checks if
/proc/$pid/status exists while opening it with fopen(). Then it reads
its lines with getline(), extracts the process state (line starting
with "State:\t"), and returns an enum value for this. The method has
proven working.
So take it over without checking the process name and wrap the !=
PROC_RUNNING check in sm_process_is_dead(). Detect the dead process
in the "set" handler in continuous mode after sleeping for one second.
In this case, set the target PID to 0 to avoid error messages when
executing "reset", inform the user about interruption, and break out
of the set loop.

Tested with Chromium B.S.U. lives, "set 0=9/1" as well as "set 0=9/5",
and terminating the game after some cycles.

Reference: #321