gui: Check target process before updating values
The values in the scan results list are set to ?? when the target
process terminates and errors are visible on the console. Those
errors are also triggered if the data_worker() is running to lock/
freeze values and to write them periodically.
So make use of sm_process_is_dead() from libscanmem, and check if
the process is running before doing the data_worker() or the
update_scan_result() tasks. The scan results list and the value
list stay at the last known values this way.

Reference: #321
Reported-by: @BluecoatEngineer