.travis.yml: Use docker and Ubuntu Bionic for gcc-7
The gcc option "-fsanitize=address,undefined" requires at least gcc-6
but Travis-CI only provides Ubuntu Trusty environments. The hack to
use a testing toolchain for gcc-6 and binutils 2.26 does not work
any more.
So use docker with the ubuntu:bionic image for the regular build.
It provides gcc-7 as the default gcc. Pass the git directory to the
Docker container and set it as the working directory.

Reference: #319