gui: Handle empty process name ValueError exception
If a running application uses an empty process name
(argv[0][0] = '\0';), then an unhandled ValueError exception is
triggered when opening the process selection dialog.
The thing which can never be empty is the path to the executable.
So gather that one from the /proc/$pid/exe symlink to handle the
exception gracefully.

Reference: #328
Suggested-by: Thomas Hess (@luziferius)