maps: Get stack end as stack load address
Using the start of the stack region as load address does not provide
a useful match offset. The actual stack end is randomized within the
upper half of the stack region. But it can be read from
/proc/$pid/stat as decimal long unsigned at the 28th value (see
"man 5 proc").
So read it with fopen(), getline(), strtoul(), and a STRCHR() macro
based on a for-loop to avoid code duplication and actual calls to
strchr(). The issue with strchr() is that it does not have a
built-in and is slow this way. Handle errors gracefully and just
return the region start in case getting the stack end fails. Check
if the stack end is really a valid value within the stack region
bounds. Only support stacks growing with lower addresses for now.