handlers: list: Show proper forward stack offset
The game endless-sky has the credits value on the stack at around
0x640 forward stack offset. Ugtrain is able to freeze it and scanmem
should be able to discover it.
So add a check to the match offset calculation: If the matching
region is a stack region and the load address is greater than the
region start, then subtract the match address from the load address
as the load address must be the stack end then.