maps: Replace alloca() with variable length array
When calling alloca() for the file name in the getline() loop, then
this allocates more and more memory on the stack for all file names
within the same maps file. It has no clue about the scope/block for
which it should keep the memory. It keeps it for the whole
sm_readmaps() function. With too many memory regions, this crashes.

So replace this with a variable length array as this one respects
the scope/block it is running in, restores the stack pointer at the
start of each loop cycle and only keeps one file name on the stack
at once this way.

Suggested-by: Andrea Stacchiotti (@12345ieee)
Reported-by: CJ Kucera (@apocalyptech)
Fixes #368