Introduce no ptrace option and write support for /proc/mem
This option permits to avoid the usage of ptrace and
to use /proc/pid/mem to read/write to memory instead for newer kernels.

Since linux kernel 3.2.2 it is now possible to read/write to memory without ptrace,
which is useful for debugging with ptrace with another process,
at the same time we use scanmem on the targeted process.

Related: #325 #371